[audio][tests] Direct-to-driver audio tests
This CL adds a new test binary audio_driver_tests that communicates
directly with audio drivers, in place of the audio_core service.
This is a significant step toward validating the compliance and
reliability of drivers, as new scenarios are enabled.

Audio drivers on the following hardware platforms have been
verified with this new test suite: NUC (HDA), Astro, Sherlock.

Test: test-only, audio_driver_tests contains 34 test cases

MTWN-328 #done

Change-Id: I178c3076f521d15e33dcb4c224b706f7aa9bb4b8