[netdump] Netemul integration tests
This change introduces integration testing to netdump by the use of
netemul virtualized networks. The code is in two parts:
- A helper library that will be shared between all netdump integration
  tests. The library sets up the test environment and organizes the
  concurrent execution of packet capture and transmission during tests.
- Test case code. This CL gives example tests for basic packet
  capture functionality using IPv4/UDP traffic generated by netstack,
  including instances where capture should fail due to bad arguments.
  Unit tests for the helper library are also included.

The test cases intentionally do not have any ad hoc timeout failure.
Instead, they rely on the global timeout in the test cmx files for
netemul to kill the test.

A mini-bug in netdump code is also fixed, after being discovered by
these tests. When an invalid capture device is supplied to netdump,
it tries to print an error message echoing the capture device path.
This sometimes comes out as garbage as the C-string pointer to
the device path is invalidated before the time the error is printed.

In forthcoming CLs there will be more tests implemented as separate
test binaries, including:
- Pcapng format test.
- Packet filter integration test.

Bug: CONN-158,
     CONN-164 #done,
     CONN-167 #done
Test: Built for core.x64 `--with //src/connectivity:tests`.
      Tested with `fx run-test netdump_integration_tests`.
      On CQ the tests will only run on NUC environments.
      This is to reduce flakiness compared to QEMU, see ZX-3348.

MULTIPLY: `[
  {
    "target": "//src/connectivity/network/netdump/integration:helper_lib_test",
    "total_runs": 200
  },
  {
    "target": "//src/connectivity/network/netdump/integration:basic_integration_test",
    "total_runs": 200
  }
]`

Change-Id: I31b921b88362c9e191a61ef2880d0380294ff2de