[bootsvc] Enable automation of integration tests
This change modifies bootsvc-integration-test to conform to the zbi_test
interface:

* when the tests pass, the test program needs to print the ZBI success
  string as described in zbi_test.gni so that infra will consider the
  test to have passed.
* when the tests have finished running, the test needs to power off the
  system so that infra does not consider the test to be hanging
  indefinitely

This covers both changes, so this change also reenables the zbi_test in
the build, which causes the zbi_tests jobs to run these tests
automatically in CQ.

In addition, I discovered that the kernel does not support the ZBI
crashlog item type on non-pc platforms.  So this change #ifdef's out the
related test assertions on non-pc platforms, since the feature is not
expected to exist there.

BUG: ZX-4844 #done
Change-Id: I723d81b99d2f2631f8d8df7d68ae34dfef1ae502