[media] Add stream failed error code
In some situations, it is of value to provide an error code pertaining
to why a Stream might have failed. This CL is the beginning of defining
and sending such error codes through the OnStreamFailed event. It does
not cover all possible error codes in all codec adapters, but it lays
the groundwork for adding them. This change will be done as a soft
transition.

MTWN-411 #comment

Change-Id: I2604fa96c0a032399cc07b1ad949a20f8256742a