[syscalls][fidl][kazoo] New backend to generate syscall files from FIDL IR
This is the beginnings of a backend for fidlc to be used to generate
syscall bindings that are currently the responsibility of abigen.

Only 2 of the simplest outputs are generated at the moment, ktrace.inc,
and kernel-branches.S. These two formats are just the names of the
syscalls with their number of arguments in two slightly different
styles.

Normal unittests for the current functionality (IR loading, the two
outputs) is added.

Additionally, //tools/kazoo/difftest can be manually run to compare
kazoo's output to banjo+abigen's output from the main build. This
difftest won't be integrated into the main build to avoid requiring
people to change both syscalls.banjo and the new fidl definitions, while
work progresses on kazoo.

ZX-3882 #comment [syscalls][fidl][kazoo] Backend to generate syscall files from FIDL IR
Bug: 33668

Change-Id: I949bd9b11410b638630a35f88176b72c9bfc16a1