[topology] On x64, fallback to basic topology if generation fails.
On x64 we auto generate the topology from cpuid and acpi information. In
some cases, for example hypervisors, this data can be inaccurate.

If we cannot properly parse an apic_id to generate the topology
structure than we fall back to pretending a flat topology with a bunch
of discrete cores.

If we can't generate a topology at all than we fallback to running with
just the boot-core in isolation, because we know that is up. This is
something we were already doing on ARM.

Test: Unit tests for the fallback to flat topology, manual test of
fallback to single core on qemu.
Issue: ZX-4348

Change-Id: I46373699bf59a834adfeef8539968342d258cc90