[media][audio_effects] Update the audio_effects ABI.
Remove the structured controls in lieu of simpler configuration blobs
which is a closer match for our current needs of the ABI. Also migrate from
per-function exports from the loadable module to exposing a single struct
that holds function pointers for each ABI method to make revisions on
the ABI simpler in the future.

Additionally, we remove the 'reset' API since as we no longer have a
concept of a 'default' configuration, this API would be identical to
flush.

We also update naming of symbols to replace 'audio_dfx' (Audio Device
Effects) with 'audio_effects'.

Test: fx run-test audio_dfx_tests
      fx run-test effects_loader_unittests
Change-Id: I1b4a673faa37a0e5f5c597b1d479e7a6053e05c8