[OpenThread] OT radio driver for Sherlock
Adding a driver for openthread radio, nrf52 and configuring reset and
interrupt GPIOs. At the moment, the driver starts up, resets the radio
and waits for interrupt from the radio to fire.

Tests: Added nrf52840-radio to device-enumeration test:
runtests -t device-enumeration-test
Verified using `dm dump` that [nrf52840_radio] and [ot-radio] are running.
Verified using logic analyzer that the radio reset is getting triggered
when the ot-radio driver starts and that the radio sends an interrupt
right after the reset.

Change-Id: I7fc74eecf5bc4ef27ea841becbca29fe633b69be