[optee] TEE Controller / Sysmem connectivity
This changes TEE Controller to a composite device that binds to sysmem,
to establish a banjo protocol connection from TEE Controller to sysmem.

This will be used to tell sysmem about VDEC memory and to permit sysmem
to protect non-VDEC memory.

Bug 34800 is opened re. adding optee_smoke_test as integration test when
possible.

TEST: fx run-test optee_smoke_test (on astro; no special setup required)
TEST: fx run-test optee_test (godtamit@ or dustingreen@ for setup)
Change-Id: Ic4d89d065477ce72fe0dd72bed5718e36aa0cd5d