[fidl][llcpp] Simple accounting for future size increase in flexible messages
When receiving a message, i.e.
  - Event handler
  - Reading response during a two-way client call
  - Reading request at the server side
one has to allocate the maximum message size for the transport, if the
message contains a type with flexible envelopes (flexible xunion or
flexible table).

Thoroughly addressing this would require making the EncodedMessage
types aware of the context: are we sending, or are we receiving.
As a first step, we assume the maximum regardless of context, if a
message contains a flexible envelope, to unblock LLCPP Tables.

Test: Added flexible_test, which tests that receiving unknown xunion
variants with more bytes/handles is ok.

FIDL-771 #comment

Change-Id: Ia00439dc67a97cdd5711b396ac3a30ac1ca4269d