[mdns] switch mDNS back to standard address/port
Prior to this CL, mDNS has been running on a non-standard port and
multicast address. This was done because port and address sharing
weren't working previously, causing the fuchsia mDNS implementation
to interfere with the operation of other mDNS implementations.

This CL makes mDNS use the standard port/address. This CL should not
be submitted until we're sure that address and port sharing work
properly.

CONN-78 #done

Change-Id: Ia918d24602c29859faccbb0da98e4d47b1f562eb