[sdk/inspect] Don't zero blocks when extending
VMOs are mapped to the zero page and are copy-on-write; all Block
structures will therefore be implicitly zeroed and the zeroing is
unnecessary.

Although I expected this to have some performance impact, it turns out
it doesn't:

Pre:
      Mean    Std dev        Min        Max     Median Unit         Mean Mbytes/sec Test case
      3951       2625       3213      56806       3504 nanoseconds              N/A Inspect/Heap/Extend.Create 1MB VMO
    315285     108827     285313    2575585     293327 nanoseconds              N/A Inspect/Heap/Extend.Allocate 512K
    292925      94310     265260    1595826     273041 nanoseconds              N/A Inspect/Heap/Extend.Extend
     29269      13474      26101     391592      27205 nanoseconds              N/A Inspect/Heap/Extend.Free
     11022       5744       9611     129755       9967 nanoseconds              N/A Inspect/Heap/Extend.Destroy

Post:
      Mean    Std dev        Min        Max     Median Unit         Mean Mbytes/sec Test case
      3874       1324       3365      31872       3651 nanoseconds              N/A Inspect/Heap/Extend.Create 1MB VMO
    301042      61845     286266    1142805     288853 nanoseconds              N/A Inspect/Heap/Extend.Allocate 512K
    279992      62414     266791    1316519     267933 nanoseconds              N/A Inspect/Heap/Extend.Extend
     27776       4864      26280      92270      26628 nanoseconds              N/A Inspect/Heap/Extend.Free
     10677       2386       9787      42867      10251 nanoseconds              N/A Inspect/Heap/Extend.Destroy

These benchmarks are noisy whether run on qemu or real hardware. An
analysis of the changes in the binary shows that this results in the
removal of two mov instructions on x86-64. These instructions are likely
nicely pipelined and executing them is likely about half a cycle. Page
faults are possible (likely?) during this process, as well as copying
when the pages are touched (which they are anyway, as the block header
is written to regardless of whether the block is zeroed). Both of these
likely dominate the runtime, and half a cycle of latency probably can't
be observed appreciably, especially since the process has high
variability anyway.

The included benchmark will hopefully still be useful for any future
improvements.

Test: benchmark passes; existing tests in CQ have good coverage.

Change-Id: I27084bb0e69c547246f0c1fee3a698f813853b5f