[netstack3] ForwardingTable to do at max 1 lookup
This change supports the seperation of a forwarding table and a more verbose/full
routing table. The routing table will store all the learned routes and install them
in the forwarding table when appropriate. This change also guarantees that when doing
a lookup for a route to a destination, the lookup depth will be at max 1.

Tests: Tests to make sure that the forwarding table only stores the best entries with a lookup
depth of at max 1.

Change-Id: Ideed38a65e33f20ea6f22624c60565d17d6b2e07