[debugger] Find std::string symbols.
The definition of std::string is in libc++ rather than in each library
that uses it. This is different than other symbols so there were some
bugs. This fixes two bugs:

👉 Adds the ability for FindName to only find actual definitions. The
index for a module will contain the definition if it exists, but only a
forward declaration if it doesn't exist in that module. This means that
the symbol search won't fall back to finding it in other modules if
there's no definition in the current one.

👉 Makes collection member extraction look up the concrete type for a
member to get the size. In all other cases the compiler will reference
a real type (not a forward declaration) because the size must be known
at the time the structure is compiled. But the extern-typedeffy-ness of
std::string means that this information is stripped so we need to
resolve the actual concrete type to get the size.

The second change required that the EvalContext be passed in more
places. cast.cc was one of the callers that previously only had an
EvalContext pointer to save from refcounting it unnecessarily. This
changes that code to use a const ref to a RefPtr which avoids the
refcount more naturally. I plan to standardize on this usage in a
followup.

Change-Id: Ic436a2cf8fdaa5cd173bb1260001ad7a6212fbf5