[netstack3] Enable/disable forwarding on a per-interface basis
Enable/Disable forwarding of IP packets on a per-interface basis.
Note, both the netstack AND the device has to have forwarding enabled
to forward an IP packet that arrived at a device

Tests: Unittest test to make sure that an interface can be set to
enable or disable forwarding. Tests to make sure a device is considered
forwarding capable only when both it and the netstack has forwarding
enabled. Tests to make sure that host->router and router->host changes
happen (e.g. stop/start NDP router solicitations).

Change-Id: Ibbff7fbd2fcf173bf73e6a699cb08dbf96645df0