[fdio] add fdio_get_vmo_exec, for requesting executable VMOs
Today all callers of fdio_get_vmo_clone explicitly add executability by
calling zx_vmo_replace_as_executable.  We'd like to make that not the
case, but to do so we'll need to migrate the callsites to something that
explicitly requests executability from the filesystem, then plumb that
bit through all the filesystem layers, and then give the ability to mark
VMOs as executable to just the trusted filesystems.

Then, we can remove calls to zx_vmo_replace_as_executable for callers
that migrate to these calls.

Change-Id: I99da6122bf2ee6469e3b935bffa0c86fee981c0d