[activity-service] Implement fuchsia.ui.activity.Tracker
fuchsia.ui.activity.Tracker is a FIDL interface which receives evidence
of user activity and uses this evidence to drive an internal state
machine to determine the idle state of the system.

This change adds a component which provides the
fuchsia.ui.activity.Tracker interface. Without the
fuchsia.ui.activity.Provider interface this is a write-only service; the
Provider interface (to be added later) will allow other components to
subscribe to state changes.

Tested: Unit tests; component startup test
Change-Id: Ibdcd64488c427e21ef8135abb8556ce631b1951e