[virtcon] Add multi display unit testing
ZX-4343 #done

Add a framework to test the rebinding and
unbinding of displays. This verifies the
multi-display flow by checking that images
and layers are created and destroyed correctly.

The FIDL calls to the display-controller are all
stubbed out, so they are assumed to function
correctly in these unit tests.

TEST: runtests -t virtual-console-display-test

Change-Id: I35c10b3656a3038a6f9e3ec9ec2166993e681dbf