[virtio] Stop using device read/write APIs
This migrates the console to directly providing a full fuchsia.io.File
interface.  Due to current devhost limitations, we need a shim interface
(fuchsia.hardware.virtioconsole.Device) to accomplish this.

The hypervisor tests use virtio-console to talk to a Fuchsia hypervisor
by way of the kernel "console.path" argument.  To keep this working,
an additional argument was added for "console.is_virtio".  Once that
devhost limitation is removed, we can remove "console.is_virtio".

Bug: 34622 #done

Change-Id: I21d2870550f3bc480f85f26ccf6e352b424ead7f