[fidlc] Add diagnostics to clashing_ordinal_values.
The clashing_ordinal_values test in ordinals_test.cc is flaking; see
FLK-435 for more details. This CL adds some diagnostics to the test to
attempt to root cause ths issue if it flakes again.

This CL is based off
<https://fuchsia-review.googlesource.com/c/fuchsia/+/298593>, but with
only changes to the ordinal_tests.cc file.

Test: fx build zircon/system/utest:host && out/default.zircon/host-x64-linux-asan/obj/system/utest/fidl-compiler/fidl-compiler-test --case protocol_tests

FLK-435 #progress

Change-Id: I8e462e8bd5f35d7c16229f29bdced8cc14e724af