[fidl] Add .Type.Identifier to C++ IR
C++ has separate namespaces for classes/structs, unions, and enums, so
"class/struct Foo;", "union Foo;" and "enum Foo;" can all exist. If all
those declarations exist, the identifier must be prefixed with
class/struct/union/enum to disambiguate which type is being referred to.

This was causing problems when declaring "struct Empty {};" in a FIDL
file, and using it in a xunion with the HLCPP bindings, since the xunion
also declares an Empty enum that's scoped to the class. That caused the
build to break with the following error:

  fidling/gen/garnet/lib/wlan/fidl/fuchsia/wlan/sme/cpp/fidl.cc:184:27: error: must use 'class' tag to refer to type 'Empty' in this scope
        new (&value->none_) Empty();
                            ^
                            class
  fidling/gen/garnet/lib/wlan/fidl/fuchsia/wlan/sme/cpp/fidl.h:235:5: note: class 'Empty' is hidden by a non-type declaration of 'Empty' here
      Empty = kUnknown,  // DEPRECATED: use kUnknown instead.
      ^

To solve this, we add a Type.Identifier() method that's similar to
Type.Decl. Type.Identifier() prefixes the type name with "class" to
disambiguate it from other declared types. All references to the type
in the C++ templates should use .Type.Identifier instead of .Type.Decl.

Currently, only the _xunion_ C++ templates use this, since the logic in
.Type.Identifier() to determine whether the class/struct type qualifier
is syntactically valid is very loose, and may not be 100% correct. If
this CL doesn't break the build, we can consider using .Type.Identifier
everywhere.

A better long-term solution is to refactor the C++ IR to expose the
fully-qualified name. That involves a lot more work, and we'd like to
unblock folks affected by this bug for the moment.

Test: fx set smart_display.astro --with //garnet/lib/wlan/fidl:service && fx clean && fx build
Test: fx exec $FUCHSIA_DIR/garnet/go/src/fidl/compiler/backend/typestest/regen.sh
Test: fx run-host-tests fidlgen_cpp_ir_test

FIDL-762 #progress

Change-Id: I5f25b7c05cb85b64f10d6964550f056575b3caf4