[pkg_resolver] Add integration test
This change adds a new integration test that verifies, end to end, that
the package resolver can cache a package from a remote TUF repository to
pkgfs.

Additionally, this change verifies that if the DOWNLOAD_BLOBS experiment
is enabled, packages fail to resolve, as that feature is not yet
implemented.

Bug: PKG-693 #comment add integration test
Change-Id: I3939eb5accaa9f84d3ea60046e91e1a4d9b3890c