[kernel][object] Always Semaphore::Wait in PortDispatcher::Dequeue
This change improves the efficiency of zx_port_wait. It does not
change any observable functionality.

When dequeuing a packet, always Wait on the Semaphore before checking
the queue to ensure that the Semaphore count does not get way out of
sync with the number of queued packets.

Prior to this change, Wait was only called when the queue was empty.
As a result it was possible for a thread to repeatedly "spin" on the
Semaphore in Dequeue (Wait returns ZX_OK, but queue is still empty;
repeat).

Note, it is still possible for the queue size to be out of sync with
the Semaphore count because queued packets may be removed via
RemoveInterruptPacket or CancelQueued.

Bug: 34636 #comment always wait
Test: ran /boot/test/core/port-test in a loop for several hours
Change-Id: I8c6441abfb72e0dcac70dd1434f116d76fcddfe2