[wlan][iwlwifi] improvement for header inclusion.
fuchsia_porting.h contains the 'glue' structs and functions when
we are porting Intel WiFi driver to fuchsia land.

However, sometimes we had to put it as the first #include in the file
to make code compiler. But it is not the best solution given that it
requires the clang exception.

So, here is the general rule to fix this issue:
+ Every .h file should be responsible for the header files it needs.
  It should include what it needs, including the fuchsia_porting.h.
+ A .c* file should not include a header, including fuchsia_porting.h,
  in order to satisfy the dependencies of a different header file.
  In this case, the include should be moved to the header file.

BUG=WLAN-1215
TEST=Build and run on pixel. Everything works fine as before.

Change-Id: I336ea5b0cd50c9e0804c54593036f90a331f1eea