[media][audio_effects] Add EffectsModule.
RAII wrapper around a loaded module that handles unloading the library
when all references to the module are released. This lays the
groundwork to enable us to  ensure that loaded effects can retain a
reference to the underlying module they're loaded from and decouple them
from the lifetime of the EffectsLoader object.

Test: fx run-test effects_loader_unittests
Change-Id: Ia81b4e876802a2a9d40829dd3b0bb782fc1133d7