[gidl] On gidl regen failure, don't write partial generated files
Pre this CL, if you are working on a GIDL feature and there is a
failure at a step in generation, you need to go through and clean
up several repos before you can build a new version of GIDL which
is a bit laborious.  This makes things easier.

Test: ran regen with failure injected in the script, ran regen with
success

Change-Id: I23d204170c0fd20cbd72b939a65c5c00275b79d3