[kernel][object][port] Unlock before posting
This change simplifies the port enqueue methods by removing the need
for AutoReschedDisable. It does not change any observable
functionality.

PortDispatcher's packet queues are protected by either a spinlock or
the Dispatcher lock. When a thread dequeues a packet, it Waits on a
semaphore to be signaled, then acquires the appropriate queue lock.

Prior to this change, the enqueuing thread used AutoReschedDisable to
ensure that any woken thread is woken after it has released the queue
lock. By dropping the lock before posting we can achieve the same
result without needing AutoReschedDisable.

Bug: 34636 #comment unlock then post
Test: ran /boot/test/core/port-test in a loop for several hours
Change-Id: Idb65075956970e1910aebb3b98e19496a3747aa6