[media][audio_effects] Add Effect class to manage effect handles.
The Effect class performs a few basic roles:
  * Takes ownership of a fuchsia_audio_effect_handle_t and ensures it
    is deleted when the Effect is destructed.
  * Retains a reference to the backing loadable module to ensure it's
    not unloaded while the Effect is active.
  * Provides access to the instance-scoped ABI methods (ex:
    process/process_inplace, etc)

Effects are move-only to ensure clear ownership of the
fuchsia_audio_effect_handle_t is maintained. After an Effect has been moved,
it will be left in an invalid state (ex: FUCHSIA_AUDIO_EFFECTS_INVALID_HANDLE)
and will cause an abort if any of the ABI methods are used.

Test: fx run-test effects_loader_unittests
Change-Id: Ie7c5882eee9e6e3bf02536a40eab835d300833f6