[debugger] Pretty-print std::optional and fit::optional
Adds a PrettyOptional type which expresses the general concept of an
optional type. This allows easy expression of pretty-printers for std::
and fit::optional types.

Adds a new FormatNode type that allows a wrapped type like an optional
to be expressed more clearly.

Remove the context parameter from a helper in pretty_tree.cc, it was
unused.

Change-Id: I842dc34465794902f2da46da89668d15781cc4c4