[syscalls][kazoo] Add backends for x86 and arm asm
These two are only slightly more complicated than the previous ones:
they include the number of arguments to the syscall which requires
munging the request/response structs (fidl's name for input and output
parameters) to determine the number of arguments that are in the C
syscall.

There's currently no difference between the x86 and arm output, so the
same backend is used, but the separate command line arguments are
maintained in case they need to diverge in the future.

Bug: 33668

Change-Id: Ibe66243b557d5a3f5d35a6e31e28ca6e1975ffd6