[kernel][vmar] Fix bug with mapping ranges with an offset
If zx_vmar_map was called with:
- A provided VMO
- An offset within the VMO to map
- ZX_VM_MAP_RANGE set as a flag

Then the vmar mapping could incorrectly use the vmo offset as the offset within
the range to map pages to. This failed size checks and caused the call to fail.
An easy way to hit this is by creating an MmioBuffer that maps a range offset
from the base of a VMO.

A test has been added to check for this.

Change-Id: I699312c93ddc76e99b7ff77901ff335a9f82613d