[a11y] Implements Gesture Manager.
This change implements Gesture Manager, used inside Accessibility Manager to
listen for incoming Pointer Events and detect gestures on a touch screen device.

This is an initial and simple implementation that only adds the possibility for
an assistive technology to bind an action to a gesture, which will be called
when the gesture is detected.

The only gesture supported for now is a one finger tap on the screen.

This serves as the basis for a more complex system, where new
gestures (specially time-based ones), will be added in a future CL.

TEST:
  fx run-test a11y_lib_tests

Change-Id: I917345f1d618e55a66012c8f0911652f4d46aeea