[netstack/dhcp] complete transition from Subnet
This was started in 6a988c3 but did not extend to the DHCP client. This
change replaces uses of tcpip.Address,tcpip.Subnet pairs with
tcpip.AddressWithPrefix, which more accurately models the types used.

As a consequence of this refactoring, we now always log the address as
well as its prefix length, which might aid with DHCP flapping seen in
b/139329385.

Change-Id: I6b904dc8989f576177a870f83958e7e314446a13