[debugger] Add an asynchronous ResolveMember().
Currently this does the same thing as the synchronous version but issues
a callback. There should be no behavior change.

This is split into synchronous (for nonstatic data) and asynchronous
(for all data) versions to support resolving static data members in a
future patch. This will require a variant with a callback.

Remove some unused helper functions on PrettyType for resolving members.

Converts some EvalContext RefPtrs to const refs and documents the new
best practice.

Change-Id: Ibaa0d0f99ba5daaa9ee9b83b74b69eb507105919