[utest][zircon] increase watchdog timeout for ASAN
Bug 34860 indicates that at least kernel-unittests are
now flaky because the watchdog fires. The sample set we
are able to find indicates the test complete in the
range from 20.5 to 28 seconds thus triggering the default
20 second (for small test) default timeout.

Since we only find ASAN examples we increase the default
to 40 seconds for that scenario.

Bug: 34860, 31966
Change-Id: I0b1cecbafb0352edf1e0dd1cb2dda59b861d0ed5