[iosched] Async op completion.
Handle asynchronous completion of operations.
Ops are now maintained in streams on the issued list until completed or
released. Currently release is done on the async completion caller
context. This works but in the future completions should defer to a
completion queue which will be serviced by worker threads.
Removed retain list in streams, ops now retained in the state lists
(input, issue, others in the future).

Tests:
Added tests to test async completion of ops, mixed with synchronous
completion. Added stream test to validate it closing waits until
pipeline is drained.

Change-Id: I6df11765e05f076945b4eb5e8ee1507179dc4940