[ermine] ignore deprecated method
Adds an ignore deprecated_member_use in the Scrim class.

This fixes an error related to onPointerHover which is deprecated in an
upcoming Flutter roll. This usage site was introduced in
https://fuchsia-review.googlesource.com/c/experiences/+/310148

In https://github.com/flutter/flutter/pull/36217, the pointer events
were split from the mouseEvents. more reasoning is documented in the
issue https://github.com/flutter/flutter/issues/36085.

To roll this through as a soft transition, we mark the usage as
ignores deprecation to allow the build to proceed.

Change-Id: I3d10550376c0eb3c0ef24c1cf596db31f1f97e4c