[debugger] Make ModuleSymbols reference counted.
Previously this class was owned by a unique_ptr and had a rather
complicated "SystemSymbols::ModuleRef" thing to manage the fact that
there may be multiple consumers referencing the object.

This change simplifies that code and also makes it possible from symbols
to the module that created them in the future.

There should be no behavior change.

Change-Id: Idf89e6a59c7c5b2dd3cdea0350884a035a8a6ef0