[media] Add error code to OnStreamFailed
In some situations, it is of value to provide an error code pertaining
to why a Stream might have failed. This CL updates OnStreamFailed to
include a StreamError error code. This is the second stage of the soft
transition and OnStreamFailed2 will remain until all clients have been
migrated to the new OnStreamFailed event.

Bug: 13691
Test: fx run-test stream_processor_lib_test
Test: fx run-test use_h264_decoder_test
Test: fx run-test audio_encoder_test

Change-Id: I13a5afba685a2f6c13ea9d1098f19d74dfb70c93