[virtualization][tests] Detect guest network interface at runtime
In Linux, network interfaces are assigned predictable names based on
their hardware. For virtio-net devices this means they get a name based
on the PCI device and which slot the device is. Unforunately the slot
number may change depending on which devices are added to the guest.

This change uses the test util, along with the guest's MAC address, to
detect the interface name after booting the guest.

Change-Id: I78a83ed598b1c3ca35a490bbd1c3b8e25856f84b