[brcmfmac] Fix ownership in the bus interface
This commit moves the state held in the brcmf_device struct into a
Device class, which proper C++ lifetime semantics.  Most of the state
formerly held in brcmf_device (which is not much) is moved into
brcmf_pub.

This allows us to fix the brcmf_bus_ops interface to take a brcmf_bus
instance as its context, rather than a brcmf_device.

This change also performs some cleanup of static state:
* The async dispatcher is now a member of Device, rather than a static
  instance, and its lifetime is properly tied to its Device.
* The irq_callback_lock is now a brcmf_pub member, and instantiated
  per-device.

Change-Id: I70362e57aa5f5aa269cc1dff919fc99c1333c468