[debugger] Resolve extern variables
Implements reading extern global variables and static structure members
defined externally. This requires converting the extern value
declaration to a real variable to get its location, and then evaluating
that.

This requires a new member is added to Value that reflects the
DW_AT_external DWARF tag. Symbol reading is updated accordingly.

This function is added to EvalContext::GetVariableValue() which handles
the global extern variable case. This function is changed to take a
Valur and not a Variable so both Variables and static DataMembers can be
passed to it.

A special code path is added to resolving structure members that detects
extern static members and resolve the variable using the above function.

Array symbols are updated to allow unknown array sizes. These can be
encountered when dealing with extern data. There is not much to do with
this type, but we can at least represent it without errors.
Documentation is added to array_type.h on this case.

Skips printing static data members when printing structs. This decision
is documented in format.cc. We can reevaluate and add in the future if
desired.

Change-Id: I65a7dda98664f8fccb1e943cfb9420fcdf885654