[debugger] Pass EvalContext by const ref.
Updates functions that take an fxl::RefPtr<EvalContext> to instead take
a const fxl::RefPtr<EvalContext>& as recommended in the eval_context.h
header.

This is because there are many functions that don't typically need to
take a reference, and our threadsafe refcounting requires a synchronized
math operation for every function call. My continuing to pass a RefPtr
we ensure that captures in lambdas still take a reference automatically.

Update some users of callback<void(ErrOrValue)> to EvalCallback.

There should be no behavior change.

Change-Id: Id5322903441d6e0641298f255b9eadf68463f4ce