[fidlc] Hack to rename fuchsia.io to fuchsia.io1
In order to make room for the new fuchsia.io library (dubbed fuchsia.io2
currently), we are piggybacking the rename of the currently named
`fuchsia.io` library to `fuchsia.io1` with the migration to 64b ordinals.
In short, from an ABI standpoint, both `fuchsia.io` and `fuchsia.io1` are
the same in the soon to be 64b ordinal world.

This will only take effect once we use 64b ordinals. It has no
impact today.

Test: ordinals_tests.cc
Change-Id: Ib7282e4162d16eea059c9b5eb2e2b712e775d6a8