[audio] Add Audio Admin Framework
This introduces the concept of audio administration to audio core.  The
first step is to extend audio core to report when streams start and
stop, and to handle some basic interactions between active streams.

A default policy is added based on behaviors seen on other systems, but
this policy can be overriden on a platform basis, or via fidl.

Test: fx build, pave
      fx run-test audio_fidl_tests
      fx run-test audio_mixer_tests
      fx run-test audio_admin_tests
      fx run-test mediaplayer_tests
      fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
      fx run-test audio_loopback_tests
Change-Id: Ib68dbab2a1b22e005bae87e12c07e2a96ebdb161