[netdump] Netdump filter integration
This CL is part of a series that adds packet filtering capability
to netdump using a filter language with similar syntax to the Berkeley
Packet Filter (BPF).

The completed filter components are integrated into the main netdump
program. In addition, a capture timeout feature is added, and packet
length checks are changed to rely upon those implemented in
fxr/299857. For packets that have missing, incomplete or unrecognized
headers in various layers, as much information about the packet
as possible is displayed.

Bug: 4864, 4894, 4906
Test:
 - Adds integration tests with netemul
   fx run-test netdump_integration_tests
   Build --with //src/connectivity:tests
 - fx run-test netdump_unit_tests
   Build --with //src/connectivity/network/netdump:netdump_unit_tests
 - Manual inspection with various filter strings

MULTIPLY: `[
  {
    "target": "//src/connectivity/network/netdump/integration:filter_integration_test",
    "total_runs": 200
  }
]`

Change-Id: If84ed1a2b4074f8a1985852a1ae9f802c3857255