[audio_core] Squelch glitch-logging
Detecting audio glitches (underflows leading to discontinuities) is a
good thing. However, logging every detected underflow -- when it is
likely that high resource contention caused it in the first place --
could reasonably be called an anti-pattern.

When the mixer entirely or partially skips a client-submitted packet,
because it considers that packet's (explicit or implicit) timestamp to
be LATE, we call that a "client-side underflow".
With this CL, audio_core tracks the number of client-side underflows,
per Renderer, with separate counts of full (whole-packet) vs. partial.
Although we tally them all, we do not log-display them all. With this
CL, we continue to log the first client "glitch" detected, but
subsequently throttle how frequently these are logged, based on
constants associated with audio_core's current logging level.

By default on non-debug builds, audio_core will log each renderer's
glitch #1, 101, 201, etc. On debug builds, audio_core will log #1, 11,
21, etc.

This CL also disables the logging of normal position notifications
altogether. Previously, these were throttled to approx. 1-3 per hour
on DEBUG and not at all on NDEBUG.

Bug: 34866
Test: logging-only

Change-Id: I4d6de50176ba192e9db6fbd7ba014de600da2a9f