[dev][usb-function] Fix test case where device is released.
If we close the cachecontrol_ device handle during the disconnect
in the ums-test, this will trigger the release function of the
ums-function device.

In ums-function bind, it looks like the writeback_cache_ check
will never be true, since we just set it to false at the start of bind.
This means we are always using the static vmo for ums->storage.

In ums-function release, we check for whether writeback_cache_ is true,
to figure out if we should release the child vmo. However, this is
actually releasing the static vmo, as we never made the copy in bind.

Next time ums-function binds again, vmar_map will fail since the static
vmo has been released.

TEST= runtests -t usb-virtual-bus-ums-test
Updated test fails before ums-function changes, passes with changes.

Change-Id: I32939cface3887c9c897a608331e1823b502db9d