[scenic] Remove NodeParts
Delete all scenic code related to node parts. All non-scenic
calls to AddParts have been removed in other CLs. Existing API
in commands.fidl union remains as to not break the ABI contract,
however it is now a no-op and will result in a session error
if an attempt is made to call it.

Testing - Added new unit test to scenic_unittest.cc to make sure
the scenic session is closed when AddParts is called.

Change-Id: Icb9332b30df4d34895c3c442791ffb441f9080c6