[fallback-rtc] Add integration tests for the fallback-rtc driver.
ZX-3816 #comment

The test is using the IsolatedDevmgr so that it can be run even on
boards that don't use this driver.
To achieve proper isolation (and have
the driver not call syscalls) a metadata is passed from the test to
indicated that the driver should operate in a sandbox mode. This is not
ideal and eventually we need to find an approach where the syscalled are
not called directly (e.g. either abstract through a lib or pass a
"testing VDSO" to the driver process).

To run the test first build a configuration with zircon tests enabled:
--with-base='garnet/packages/tests:zircon'

Then pave a device and from the serial console run "runtests -t fallback-rtc"

TEST: runtests -t fallback-rtc

Change-Id: I5f63097612b30213f106a560059b969b70454c81