[audio_core] Add capture underrun detection (and logging)
To increase diagnosability, this CL adds underrun-detection logging
to the audio capturer. It also clarifies a number of related variables
and comments. Logging traffic may increase (if the system is failing
to perform as required), but no other behavior change.

Most importantly, in cases where our initial capture ring-buffer
position is calculated to be negative (because of offsets due to
hardware FIFO depth), correctly wrap this around instead of ASSERT.

Test: change in logging behavior and DEBUG error checking

Bug: 8587
Bug: 13693
Bug: 34866

Change-Id: Ic4eb9f506e52562376910c67670ebc241493ce76