[ledger] Top-level FIDL in integration tests
At least for our integration tests we now no longer need the hack of
"here's a top-level node and here's an 'attachment' node under it to
which the system under tests attaches". And good, because it was making
tests of later commits flake.

A TODO linked to
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=12278 is added for
making the same FIDL connection work in the end-to-end tests.

Change-Id: I885bbe2358d0735b2affc2a2cc6413b1e25e9a1d