[ledger] Page heads ChildrenManager integration
The HeadsChildrenManager is one of two ChildrenManager implementations
that will be held by the PageManager (the other being the yet-to-come
CommitsChildrenManager).

As for the Inspect nodes representing the heads: they report to Inspect
all that they have to report by their mere existence, so there's no
need for them to expose any properties or metrics.

Change-Id: I58cedf906b7fef8901c9f7908fbd3f22f105dd70