[zedboot][zero-state] Insert a BootZedbootShim().
This is the first step in getting to a place where hardware is delivered
to the platform in a consistent known state (Zedboot @ build revision).

Eventually we will pass the device off with the correct version of
Zedboot to the platform from infra, but this serves as an emulation
until we reach that state.

Bug: 10241
Change-Id: Iaaab821e5142075473e64fcffbd287ecb93d3023