[e2e][webdriver] Log Chromedriver errors
This change surfaces any errors from the Chromedriver binary
started by the webdriver library.  The intent of this change
is to enable debugging of errors in test bots.  The logs
output by Chromedriver are minimal - a few lines to stdout on
startup.

Testing: No change in behavior.  Verify that logs are present
when running e2e test with Chromedriver

Change-Id: If1cc6f29826d630ea0e38f14ee63356d50ffeb2f