[syscalls][kazoo] Refactoring of syscall loading
This is an attempt to break up
https://fuchsia-review.googlesource.com/c/fuchsia/+/311026 (which adds
the first backend that uses type information).

This CL is mostly concerned with creating the kernel organization of
arguments out of the fidl representation. There was a first pass at that
previously in output_util.cc, but most backends will need the kernel
version, so it makes sense to centralize it.

This mapping maps:
- from multiple inputs/outputs to C-style with a single return value and
  outparams
- from fidl higher level types like "string" and "vector" to
  pointer+sizes (along with some name futzing to match standard uses in
  our current API)

There's a test added for some of this, but the main test is that the
other generators don't break.

Bug: 33668
Change-Id: I1a701295d3cf8ec3a21da61b0be10601780d0a0a